refactor : api 훅 삭제 명 수정
